requirements.txt
boto3==1.4.6
botocore==1.6.5
docutils==0.14
futures==3.1.1
jmespath==0.9.3
python-dateutil==2.6.1
pyyaml>=4.2b1
s3transfer==0.1.10
six==1.10.0
wheel==0.24.0